Fix for non-json encoded payloads
This will make two attempts to decode the payload. If it cannot decode
the first time it will simply encode the payload then re-attempt to
decode.

While this feels like just additional overhead (why not just reply with
the original payload?) this will hopefully keep the API on the path of
using JSON responses an anticipation that the Chromium team fixes this
bug upstream. This way we don't have to introduce a bunch of forking
code to detect different payload types.